requirements.txt
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=1.0.0
nltk>=3.6.0
wikipedia>=1.4.0
matplotlib>=3.3.0
Pillow>=8.0.0
tensorflow>=2.8.0
requests>=2.25.0
scipy>=1.7.0
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.6.0 